Feature/33408 convert to systemd (#27)
* relocate files to single tree
* switch to systemd services and fix minor bugs
 * remove unused code from Makefile
 * add separated wb-gsm-rtc restore and save services (which checks if
   modem exists before running)
 * of: treat disabled nodes as non-existent
 * add lock in wb_env_cache generator to prevent simultaneous runs
   during bootup
 * wb-init: fix resetting RTS state on WB5
 * wb-prepare: check empty files in wb_fix_file
 * wb-set-mac: use ip link instead of ifconfig
 * wb-prepare: check if file is the same before copying from /mnt/data
 * wb-prepare: run before avahi-daemon

+++ utils/lib/prepare/wb-prepare.sh
@@ -7,8 +7,8 @@ ROOT_PARTITION=$(mount -l | grep " / " | cut -d " " -f1)
 . /lib/lsb/init-functions
 
 # some constants for convenience
-. /usr/lib/wb-prepare/vars.sh
-. /usr/lib/wb-prepare/partitions.sh
+. /usr/lib/wb-utils/prepare/vars.sh
+. /usr/lib/wb-utils/prepare/partitions.sh
 
 MB=1024*1024
 
@@ -203,10 +203,10 @@ wb_fix_file()
     local descr=$2
     local cmd=$3
 
-    if [[ ! -f ${file} ]]; then
+    if [[ ! -f ${file} ]] || [[ ! -s ${file} ]]; then
         log_action_begin_msg "Creating ${descr}"
 
-        if [[ -e "/mnt/data/${file}" ]]; then
+        if [[ -e "/mnt/data/${file}" ]] && [[ ! "/mnt/data/${file}" -ef "${file}" ]] ; then
             log_action_cont_msg "found at /mnt/data"
             cp "/mnt/data/${file}" "${file}"
         else
@@ -266,13 +266,13 @@ wb_fix_short_sn()
         hostnamectl set-hostname --static $hostname
         hostnamectl set-chassis embedded
     else
-        echo $hostname > /etc/hostname.wb
+        echo "$hostname" > /etc/hostname.wb
         FIRSTBOOT_NEED_REBOOT=true
     fi
     log_end_msg $?
 
     log_action_msg "Setting internal Wi-Fi SSID to ${ssid}"
-    sed -i "s/^ssid=.*/ssid=${ssid}/" `readlink -f "/etc/hostapd.conf"`
+    sed -i "s/^ssid=.*/ssid=${ssid}/" $(readlink -f "/etc/hostapd.conf")
     log_end_msg $?
 }
